+++ api/src/domain/product/aggregate_root.rs
@@ -16,7 +16,7 @@ pub struct Product {
     image_url:           Option<String>,
     low_stock_threshold: u32,
     name:                Name,
-    price_cents:         u64,
+    price_cents:         u32,
     stock:               u32,
     symbols:             Vec<Symbol>,
     unit_of_measure:     UnitOfMeasure,
@@ -28,7 +28,7 @@ impl Product {
         image_url:           Option<String>,
         low_stock_threshold: u32,
         name:                Name,
-        price_cents:         u64,
+        price_cents:         u32,
         stock:               u32,
         symbols:             Vec<Symbol>,
         unit_of_measure:     UnitOfMeasure,
@@ -57,7 +57,7 @@ impl Product {
         }
     }
 
-    pub fn price_for(&self, amount: f64, unit: &UnitOfMeasure) -> Result<u64, UnitOfMeasureError> {
+    pub fn price_for(&self, amount: f64, unit: &UnitOfMeasure) -> Result<u32, UnitOfMeasureError> {
         if unit.dimension() != self.unit_of_measure.dimension() {
             return Err(UnitOfMeasureError::Invalid(format!(
                 "cannot convert {:?} to {:?}",
@@ -68,7 +68,7 @@ impl Product {
         let order_base   = amount * unit.to_base_factor();
         let product_base = self.unit_of_measure.to_base_factor();
         let ratio        = order_base / product_base;
-        Ok((ratio * self.price_cents as f64).round() as u64)
+        Ok((ratio * self.price_cents as f64).round() as u32)
     }
 
     pub fn reconstitute(
@@ -77,7 +77,7 @@ impl Product {
         image_url:           Option<String>,
         low_stock_threshold: u32,
         name:                Name,
-        price_cents:         u64,
+        price_cents:         u32,
         stock:               u32,
         symbols:             Vec<Symbol>,
         unit_of_measure:     UnitOfMeasure,
@@ -102,7 +102,7 @@ impl Product {
     pub fn get_image_url(&self)           -> Option<&str>         { self.image_url.as_deref() }
     pub fn get_low_stock_threshold(&self) -> u32                  { self.low_stock_threshold }
     pub fn get_name(&self)                -> &Name                { &self.name }
-    pub fn get_price_cents(&self)         -> u64                  { self.price_cents }
+    pub fn get_price_cents(&self)         -> u32                  { self.price_cents }
     pub fn get_stock(&self)               -> u32                  { self.stock }
     pub fn get_symbols(&self)             -> &[Symbol]            { &self.symbols }
     pub fn get_unit_of_measure(&self)     -> &UnitOfMeasure       { &self.unit_of_measure }
